Rename[]: 생성 뷰모델 패키지
